requirements.txt
# aiGn Career Agent - Dependencies
# Install: pip install -r requirements.txt

# Core LLM
llama-cpp-python>=0.3.0      # For GGUF model inference
smolagents>=1.0.0            # HuggingFace agent framework
litellm>=1.0.0               # LLM API bridge (for Ollama)
mcp>=1.0.0                   # Model Context Protocol

# Document parsing
PyMuPDF>=1.24.0              # PDF text extraction (imported as fitz)
python-docx>=1.1.0           # DOCX text extraction

# Job search
python-jobspy>=1.1.0         # Indeed/Glassdoor scraper

# Data handling
pandas>=2.0.0                # DataFrame operations (for jobspy)
pydantic>=2.0.0              # Data validation

# Utilities
requests>=2.31.0             # HTTP requests
python-dotenv>=1.0.0         # Environment variables
pyperclip>=1.8.0             # Clipboard operations
prompt-toolkit>=3.0.0        # Interactive CLI input
